% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/immunogenicity_functions.R
\docType{methods}
\name{evalBoxplot}
\alias{evalBoxplot}
\alias{evalBoxplot,ImmunoAssay-method}
\title{Evaluate the Assays with Boxplots}
\usage{
evalBoxplot(assay.obj, var = c("Day", "Operator"))
}
\arguments{
\item{assay.obj}{An ImmunoAssay object imported by importAssay}

\item{var}{Variable to examine in the plot. Either "Day" or "Operator".}
}
\value{
A boxplot as generated by \code{ggplot2}
}
\description{
This function produces a boxplot based on the variable chosen in order to visualize any analytical variability.
}
\examples{
assay.obj <- importAssay(lognormAssay, exp.name = 'Experiment1')
evalBoxplot(assay.obj,var='Day') #visualize days on boxplot

}
\author{
Emma Gail
}
