% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/immunogenicity_functions.R
\docType{class}
\name{ImmunoAssay-class}
\alias{ImmunoAssay-class}
\alias{ImmunoAssay}
\title{Define ImmunoAssay class}
\description{
This stores the data that is used for screening cut point analysis.
}
\section{Slots}{

\describe{
\item{\code{data}}{Imported data as is, used for CV analysis}

\item{\code{melted.data}}{Data used for most functions}

\item{\code{exp.name}}{Experiment name}

\item{\code{stats}}{List of statistics, results gathered from both coefficient of variation analysis as well as plot generation}

\item{\code{outlier.rm}}{Has any outlier analysis been performed on this dataset?}

\item{\code{outlier.rm.method}}{If outlier removal has been performed, what method was used?}

\item{\code{scp.table}}{Table of cut point information}

\item{\code{cv.table}}{Table derived from coefficient of variation analysis}
}}

