% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/immunogenicity_functions.R
\name{importAssay}
\alias{importAssay}
\title{Import assay as ImmunoAssay object}
\usage{
importAssay(assay.df, exp.name = NULL)
}
\arguments{
\item{assay.df}{Pathname to (.csv or .xlsx files accepted) or imported data.frame consisting of the following columns: 'ID','Lot', and columns identifying the Day, Operator and Replicate like so: 'D1_Op2_3' to indicate Day 1, operator 2, replicate 3.}

\item{exp.name}{Experiment name. If stays NULL, will automatically name experiment as 'experiment1'.}
}
\value{
An object of the class ImmunoAssay
}
\description{
Function to import assay information into an ImmunoAssay object for analysis.
}
\examples{

assay.df <- importAssay(assay.df = lognormAssay)
}
\author{
Emma Gail
}
