% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/immunogenicity_functions.R
\docType{data}
\name{lognormAssay}
\alias{lognormAssay}
\title{Simulated Lognormal Dataset}
\format{
An object of class \code{data.frame} with 100 rows and 20 columns.
}
\usage{
data(lognormAssay)
}
\description{
This is a simulated dataset that using a lognormal distribution
}
\keyword{datasets}
